requirements.txt
streamlit>=1.28.0
cerebras_cloud_sdk>=1.0.0
pyperclip>=1.8.2